requirements.txt
click
azure-cosmos
azure-identity
requests